[feat] 기본 프로필 추가
