+++ SMASHING/Presentation/Notification/NotificationCell.swift
@@ -10,11 +10,7 @@ import UIKit
 import SnapKit
 import Then
 
-final class NotificationCell: BaseUICollectionViewCell, ReuseIdentifiable {
-    
-    // MARK: - Properties
-    
-    private var notification: NotificationSummaryResponseDTO?
+final class NotificationCell: UICollectionViewCell, ReuseIdentifiable {
     
     // MARK: - UI Components
     
@@ -22,7 +18,6 @@ final class NotificationCell: BaseUICollectionViewCell, ReuseIdentifiable {
         $0.contentMode = .scaleAspectFill
         $0.clipsToBounds = true
         $0.layer.cornerRadius = 20
-        $0.backgroundColor = .Background.overlay
     }
     
     private let typeLabel = UILabel().then {
@@ -42,13 +37,27 @@ final class NotificationCell: BaseUICollectionViewCell, ReuseIdentifiable {
         $0.lineBreakMode = .byWordWrapping
     }
     
-    // MARK: - Setup Methods
-    
-    override func setUI() {
-        contentView.addSubviews(profileImage, typeLabel, timeLabel, contentLabel)
+    private let dividerView = UIView().then {
+        $0.backgroundColor = .Border.primary
     }
     
-    override func setLayout() {
+    // MARK: - Setup Methods
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setUI()
+        setLayout()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func setUI() {
+        contentView.addSubviews(profileImage, typeLabel, timeLabel, contentLabel, dividerView)
+    }
+    
+    func setLayout() {
         profileImage.snp.makeConstraints {
             $0.top.equalToSuperview().inset(12)
             $0.size.equalTo(40)
@@ -72,6 +81,13 @@ final class NotificationCell: BaseUICollectionViewCell, ReuseIdentifiable {
             $0.trailing.equalToSuperview().inset(16)
             $0.bottom.equalToSuperview().inset(12)
         }
+        
+        dividerView.snp.makeConstraints {
+            $0.top.equalTo(contentLabel.snp.bottom).offset(16)
+            $0.leading.trailing.equalToSuperview()
+            $0.bottom.equalToSuperview()
+            $0.height.equalTo(1)
+        }
     }
     
     func configure(notification: NotificationSummaryResponseDTO) {
@@ -82,6 +98,7 @@ final class NotificationCell: BaseUICollectionViewCell, ReuseIdentifiable {
         }
         timeLabel.text = notification.createdAt.toDateFromISO8601?.toRelativeString()
         contentLabel.text = notification.content
+        profileImage.image = UIImage.defaultProfileImage(name: notification.senderNickname)
         backgroundColor = notification.isRead ? .clear : .Background.surface
     }
 }
